shush about the warning

+++ src/main/java/xyz/lilyflower/solaris/configuration/modules/SolarisContent.java
@@ -5,14 +5,14 @@ import java.util.Map;
 import xyz.lilyflower.solaris.api.ConfigurationModule;
 import xyz.lilyflower.solaris.configuration.SolarisConfigurationLoader;
 import xyz.lilyflower.solaris.init.Solaris;
-import xyz.lilyflower.solaris.util.SolarisExtensions.TriPair;
+import xyz.lilyflower.solaris.util.data.TriPair;
 import xyz.lilyflower.solaris.world.DimensionalWorldType;
 
 @SuppressWarnings("unused")
 public class SolarisContent implements ConfigurationModule {
     public static boolean ENABLE_CONTENT = false;
     public static String MODPACK_IDENTIFIER = "";
-    public static final Map<Integer, TriPair<Integer, Integer, Integer>> COORDINATES = new LinkedHashMap<>();
+    public static final Map<Integer, TriPair<Double, Double, Double>> COORDINATES = new LinkedHashMap<>();
 
     @Override
     public void init() {
@@ -35,9 +35,9 @@ public class SolarisContent implements ConfigurationModule {
                 if (split.length != 3) {
                     Solaris.LOGGER.warn("Coordinate pair '{}' was not exactly three long!", tuple);
                 } else {
-                    int x = Integer.parseInt(split[0]);
-                    int y = Integer.parseInt(split[1]);
-                    int z = Integer.parseInt(split[2]);
+                    double x = Double.parseDouble(split[0]);
+                    double y = Double.parseDouble(split[1]);
+                    double z = Double.parseDouble(split[2]);
                     COORDINATES.put(Integer.parseInt(dimensions[index]), new TriPair<>(x, y, z));
                 }
             }
